+++ src/test/groovy/ru/vyarus/java/generics/resolver/inlying/InlyingGenericsResolutionTest.groovy
@@ -6,6 +6,7 @@ import ru.vyarus.java.generics.resolver.inlying.support.DeclarationType
 import ru.vyarus.java.generics.resolver.inlying.support.Err
 import ru.vyarus.java.generics.resolver.inlying.support.RootType
 import ru.vyarus.java.generics.resolver.inlying.support.SubType
+import ru.vyarus.java.generics.resolver.support.Root
 import spock.lang.Specification
 
 /**
@@ -21,6 +22,12 @@ class InlyingGenericsResolutionTest extends Specification {
 
         when: "field context"
         def res = context.inlyingFieldType(DeclarationType.getDeclaredField("one"))
+        then:
+        res.generic("T") == Integer
+        res.rootContext().currentClass() == DeclarationType.class
+
+        when: "field context with interface"
+        res = context.inlyingFieldType(DeclarationType.getDeclaredField("two"))
         then:
         res.generic("T") == Integer
         res.rootContext().currentClass() == DeclarationType.class
@@ -41,12 +48,23 @@ class InlyingGenericsResolutionTest extends Specification {
         context.method(DeclarationType.getMethod("param", SubType.class)).parameterInlyingType(2)
         then:
         def ex = thrown(IllegalArgumentException)
-        ex.message == "Can't request parameter 2 of method 'param' (DeclarationType) becuase it have only 1 parameters"
+        ex.message == "Can't request parameter 2 of method 'param' (DeclarationType) because it have only 1 parameters"
 
         when: "wrong field"
         context.inlyingFieldType(Err.getDeclaredField("wrongField"))
         then: "err"
         ex = thrown(IllegalArgumentException)
         ex.message == "Field 'wrongField' declaration type Err is not present in hierarchy of RootType"
+    }
+
+    def "Check inlying type without generics"() {
+
+        setup: "prepare base type context"
+        GenericsContext context = GenericsResolver.resolve(RootType)
+
+        when: "field without generics"
+        def res = context.inlyingFieldType(RootType.getDeclaredField("nogen"))
+        then:
+        res.rootContext().currentClass() == RootType.class
     }
 }