+++ src/test/groovy/ru/vyarus/java/generics/resolver/inlying/support/NoGenericType.groovy
@@ -1,0 +1,8 @@
+package ru.vyarus.java.generics.resolver.inlying.support
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 08.05.2018
+ */
+class NoGenericType {
+}