+++ src/test/groovy/ru/vyarus/java/generics/resolver/inlying/support/RootType.groovy
@@ -5,4 +5,6 @@ package ru.vyarus.java.generics.resolver.inlying.support
  * @since 07.05.2018
  */
 class RootType extends DeclarationType<Integer, String, Double> {
+
+    NoGenericType nogen
 }